% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpreview.R
\name{ggpreview}
\alias{ggpreview}
\title{Preview plot instead of saving.}
\usage{
ggpreview(..)
}
\arguments{
\item{...}{arguments to be passed to [ggplot2::ggsave()]]}
}
\description{
Useful for checking scale of graphs
}
