% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reveal.R
\name{plot_reveal_by_group}
\alias{plot_reveal_by_group}
\title{Reveal plots by group}
\usage{
plot_reveal_by_group(plot)
}
\arguments{
\item{plot}{A ggplot2 plot}
}
\description{
Receives plot, creates steps by group and returns objects ready for saving.
Order of facet is rowise.
}
\details{
Current know limitations:
does not work with geom_smooth;
does not work with separate group aesthetics by layer
does not work if there are separate alpha aesthetics by layer
does not produce an empty first graph
}
\examples{
# Create full plot
library(ggplot2)
data("iris")

plot <- iris |>
  ggplot(aes(Sepal.Length, Sepal.Width,
             group=Species,
             color=Species)) +
  geom_point() +
  geom_line(stat="smooth",
            method = "lm") +
  theme_light()

# Create starting object
p_steps <- plot_reveal_by_group(p)

\dontrun{
# Save plots
plot_reveal_save(p_steps, "myplot")
}
}
