% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reveal.R
\name{plot_reveal_by_layer}
\alias{plot_reveal_by_layer}
\title{Reveal plots by layer}
\usage{
plot_reveal_by_layer(plot)
}
\arguments{
\item{plot}{A ggplot2 plot}
}
\description{
Receives plot, creates steps by facet and returns objects ready for saving.
Order of facet is rowise.
}
\details{
Current know limitations:
scale ranges might differ depending on which layers are included
}
\examples{
# Create full plot
library(ggplot2)
data("iris")

plot <- iris |>
  ggplot(aes(Sepal.Length, Sepal.Width,
             group=Species,
             color=Species)) +
  geom_point() +
  geom_line(stat="smooth",
            method = "lm") +
  theme_light()

# Create starting object
p_steps <- plot_reveal_by_layer(p)

\dontrun{
# Save plots
plot_reveal_save(p_steps, "myplot")
}
}
