% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reveal.R
\name{plot_reveal_save}
\alias{plot_reveal_save}
\title{Saves plot steps}
\usage{
plot_reveal_save(plot_list, basename = "plot", ...)
}
\arguments{
\item{plot_list}{The list of plots returned by `plot_reveal_steps()`}

\item{basename}{The base filename for exported plots}

\item{...}{Additional arguments (e.g. width, height) tobe passed to `ggsave`}
}
\description{
Saves plot steps
}
\examples{
# Create full plot
library(ggplot2)
data("mtcars")

p <- mtcars |>
  ggplot(aes(mpg, disp)) +
   geom_point() +
   facet_grid(vs~am) +
   theme_light()

# Create starting object
p_start <- plot_reveal_start(p)

# Examine layout and select elements to keep
steps <- list(
                c("axis-l-1" , "panel-1-1", "strip-t-1",
                  "axis-t-1",  "axis-b-1", "axis-r-1", "strip-r-1",
                  "xlab-b", "ylab-l"),

                c("axis-l-2", "panel-2-1",
                  "strip-r-2", "axis-r-2")
             )

# Create incremental plots
p_steps  <- plot_reveal_steps(p_start, steps)

# Save plots
plot_reveal_save(p_steps, "myplot")
}
